Update SkBits2Float and SkFloat2Bits to use uint32_t.
The most recent version of clang now enforces that %x must be
passed an unsigned value. SkFloat2Bits is most commonly used for
printing, so it now returns a uint32_t. For parity, its sibling
API now uses uint32_t as well. In practice, almost every call
site was already using unsigned ints, or was using bitwise math
where the signs did not matter.

Reviewed-on: https://skia-review.googlesource.com/c/skia/+/842119
Commit-Queue: Brian Osman <brianosman@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>

+++ src/base/SkRandom.h
@@ -53,7 +53,7 @@ public:
      *  Returns value [0...1) as an IEEE float
      */
     float nextF() {
-        int floatint = 0x3f800000 | (int)(this->nextU() >> 9);
+        uint32_t floatint = 0x3f800000 | (this->nextU() >> 9);
         float f = SkBits2Float(floatint) - 1.0f;
         return f;
     }
